requirements.txt
websockets
pandas
numpy
pytest
pytest-asyncio
python-dotenv
requests
